feat: Add 'Create Site Here' button for temp location, extract component from use in dashbaord
